% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{http_problem_types}
\alias{http_problem_types}
\title{List Built-In Problem Types}
\usage{
http_problem_types()
}
\value{
A data frame of HTTP status codes and their default title & type.
}
\description{
Many APIs will not need to define custom problem "types", since HTTP status
codes \href{https://tools.ietf.org/html/rfc7807#section-4}{are usually illustrative enough}.
This function lists the default type and title information for a given
status code.
}
